requirements.txt
Flask==1.1.2
imageio==2.8.0
pandas==1.0.3
numpy==1.18.4
Pillow==7.1.2
tensorflow-gpu==2.2.0